made script work with json files, without api

+++ connector/scripts/restore_database.py
@@ -7,17 +7,16 @@ class Command(BaseCommand):
     """
     Restore database from scrapper changes
     """
-    def handle(self):
-        for e in MapperEntity.objects.all():
-            e.delete()
+    def handle(self, *args, **kwargs):
 
         for team in Team.objects.filter(scrapper_autocreated=True):
+            team.mapper.delete()
             team.delete()
 
         for club in Club.objects.filter(scrapper_autocreated=True):
+            club.mapper.delete()
             club.delete()
 
-        # for team in Team.objects.filter(scrapper_teamhistory_id__isnull=False):
         for team in Team.objects.filter(mapper__isnull=False):
             entities = MapperEntity.objects.filter(target=team.mapper)
             for entity in entities:
@@ -27,16 +26,16 @@ class Command(BaseCommand):
             entities = MapperEntity.objects.filter(target=club.mapper)
             for entity in entities:
                 entity.delete()
-            # club.scrapper_uuid = None
-            # club.save()
 
         for league in League.objects.filter(scrapper_autocreated=True):
             league.delete()
 
         for lh in LeagueHistory.objects.all():
-            lh.mapper.delete()
+            if lh.mapper:
+                lh.mapper.delete()
             lh.delete()
 
         for th in TeamHistory.objects.all():
-            th.mapper.delete()
+            if th.mapper:
+                th.mapper.delete()
             th.delete()
